2D primitives were changed to doubles and SwiftProcessing global attributes were changed to doubles. Just enough were changed to accommodate beginners. This will need to be an ongoing process. Wherever possible, functions were duplicated to ensure existing SwiftProcessing code does not break.

+++ Sources/SwiftProcessing/Core/Sketch.swift
@@ -28,11 +28,11 @@ import SceneKit
     * MARK: - TRIGONOMETRY CONSTANTS
     */
 
-    public let HALF_PI = CGFloat.pi / 2
-    public let PI = CGFloat.pi
-    public let QUARTER_PI = CGFloat.pi / 4
-    public let TWO_PI = CGFloat.pi * 2
-    public let TAU = CGFloat.pi * 2
+    public let HALF_PI = Double.pi / 2
+    public let PI = Double.pi
+    public let QUARTER_PI = Double.pi / 4
+    public let TWO_PI = Double.pi * 2
+    public let TAU = Double.pi * 2
     
     /*
     * MARK: - KEYWORD CONSTANTS
@@ -95,34 +95,34 @@ import SceneKit
     */
     
     public weak var sketchDelegate: SketchDelegate?
-    public var width: CGFloat = 0
-    public var height: CGFloat = 0
-    public var nativeWidth: CGFloat = 0
-    public var nativeHeight: CGFloat = 0
-    public let deviceWidth = UIScreen.main.bounds.width
-    public let deviceHeight = UIScreen.main.bounds.height
+    public var width: Double = 0
+    public var height: Double = 0
+    public var nativeWidth: Double = 0
+    public var nativeHeight: Double = 0
+    public let deviceWidth = Double(UIScreen.main.bounds.width)
+    public let deviceHeight = Double(UIScreen.main.bounds.height)
     
     public var isFaceMode: Bool = false
     public var isFaceFill: Bool = true
     
-    public var frameCount: CGFloat = 0
-    public var deltaTime: CGFloat = 1/60
-    private var lastTime: CGFloat = CGFloat(CACurrentMediaTime())
-    var fps: CGFloat = 60
+    public var frameCount: Double = 0
+    public var deltaTime: Double = 1/60
+    private var lastTime: Double = CACurrentMediaTime()
+    var fps: Double = 60
     
     var fpsTimer: CADisplayLink?
 
-    var strokeWeight: CGFloat = 1
+    var strokeWeight: Double = 1
     var isFill: Bool = true
     var isStroke: Bool = true
     var isErase: Bool = false
     
     var isScrollX: Bool = false
     var isScrollY: Bool = true
-    var minX: CGFloat = 0
-    var maxX: CGFloat = 0
-    var minY: CGFloat = 0
-    var maxY: CGFloat = 0
+    var minX: Double = 0
+    var maxX: Double = 0
+    var minY: Double = 0
+    var maxY: Double = 0
     
     var settingsStack: SketchSettingsStack = SketchSettingsStack()
     var settings: SketchSettings = SketchSettings()
@@ -136,8 +136,8 @@ import SceneKit
     
     open var touches: [Vector] = []
     open var touched: Bool = false
-    open var touchX: CGFloat = -1
-    open var touchY: CGFloat = -1
+    open var touchX: Double = -1
+    open var touchY: Double = -1
     
     var touchMode: String = "self"
     open var SELF: String = "self"
@@ -255,10 +255,10 @@ import SceneKit
     }
     
     private func updateDimensions() {
-        self.width = self.frame.width
-        self.height = self.frame.height
-        self.nativeWidth = self.frame.width * UIScreen.main.scale
-        self.nativeHeight = self.frame.height * UIScreen.main.scale
+        self.width = Double(self.frame.width)
+        self.height = Double(self.frame.height)
+        self.nativeWidth = Double(self.frame.width) * Double(UIScreen.main.scale)
+        self.nativeHeight = Double(self.frame.height) * Double(UIScreen.main.scale)
         //recreate the backing ImageContext when the native dimensions do not match the context dimensions
         if (self.context?.width != Int(nativeWidth)
             || self.context?.height != Int(nativeHeight)) {
@@ -270,7 +270,7 @@ import SceneKit
     
     private func updateTimes() {
         frameCount =  frameCount + 1
-        let newTime = CGFloat(CACurrentMediaTime())
+        let newTime = CACurrentMediaTime()
         deltaTime = newTime - lastTime
         lastTime = newTime
     }
